requirements.txt

Django>=4.2
google-genai
python-dotenv
PyPDF2
